Generify TabDeckEditor further.
