+++ cockatrice/src/client/tabs/tab_generic_deck_editor.h
@@ -3,7 +3,10 @@
 
 #include "../../deck/custom_line_edit.h"
 #include "../../game/cards/card_database.h"
-#include "../game_logic/key_signals.h"
+#include "../ui/widgets/deck_editor/deck_editor_card_info_dock_widget.h"
+#include "../ui/widgets/deck_editor/deck_editor_deck_dock_widget.h"
+#include "../ui/widgets/deck_editor/deck_editor_filter_dock_widget.h"
+#include "../ui/widgets/deck_editor/deck_editor_printing_selector_dock_widget.h"
 #include "../ui/widgets/printing_selector/printing_selector.h"
 #include "../ui/widgets/visual_deck_storage/deck_preview/deck_preview_deck_tags_display_widget.h"
 #include "tab.h"
@@ -13,13 +16,16 @@
 
 class CardDatabaseModel;
 class CardDatabaseDisplayModel;
-class DeckListModel;
+
 class QTreeView;
 
 class CardInfoFrameWidget;
 class QTextEdit;
 class QLabel;
 class DeckLoader;
+class DeckEditorDeckDockWidget;
+class DeckEditorFilterDockWidget;
+class DeckEditorPrintingSelectorDockWidget;
 class DeckPreviewDeckTagsDisplayWidget;
 class Response;
 class FilterTreeModel;
@@ -36,11 +42,6 @@ class TabGenericDeckEditor : public Tab
 {
     Q_OBJECT
 protected slots:
-    void updateName(const QString &name);                                                         // generic
-    void updateComments();                                                                        // generic
-    void updateBannerCardComboBox();                                                              // generic
-    void setBannerCard(int);                                                                      // generic
-    void updateHash();                                                                            // generic
     void updateCardInfoLeft(const QModelIndex &current, const QModelIndex &previous);             // generic
     void updateCardInfoRight(const QModelIndex &current, const QModelIndex &previous);            // generic
     void updatePrintingSelectorDatabase(const QModelIndex &current, const QModelIndex &previous); // possibly generic
@@ -66,22 +67,13 @@ protected slots:
     void actAnalyzeDeckDeckstats();                     // generic
     void actAnalyzeDeckTappedout();                     // generic
 
-    void actClearFilterAll(); // generic
-    void actClearFilterOne(); // generic
-
-    void actSwapCard();                   // probably generic
     void actAddCard();                    // generic
     void actAddCardToSideboard();         // generic
-    void actRemoveCard();                 // generic
-    void actIncrement();                  // generic
-    void actDecrement();                  // generic
     void actDecrementCard();              // generic
     void actDecrementCardFromSideboard(); // generic
     void copyDatabaseCellContents();      // generic
 
     void saveDeckRemoteFinished(const Response &r);        // generic
-    void filterViewCustomContextMenu(const QPoint &point); // generic
-    void filterRemove(QAction *action);                    // generic
 
     virtual void loadLayout() = 0;
     virtual void restartLayout() = 0; // somewhat generic
@@ -93,7 +85,6 @@ protected slots:
     virtual void dockFloatingTriggered() = 0;            // custom
     virtual void dockTopLevelChanged(bool topLevel) = 0; // custom
     void saveDbHeaderState();                            // generic
-    void setSaveStatus(bool newStatus);                  // generic
     void showSearchSyntaxHelp();                         // generic
 
 protected:
@@ -111,39 +102,12 @@ protected:
 
     bool isBlankNewDeck() const;
     CardInfoPtr currentCardInfo() const;                         // generic
-    void offsetCountAtIndex(const QModelIndex &idx, int offset); // generic
     void decrementCardHelper(QString zoneName);                  // generic
-    bool swapCard(const QModelIndex &idx);                       // probably generic
-    void recursiveExpand(const QModelIndex &index);              // generic
     virtual void openDeckFromFile(const QString &fileName, DeckOpenLocation deckOpenLocation) = 0;
 
-    QModelIndexList getSelectedCardNodes() const;
-
     CardDatabaseModel *databaseModel;
-    CardDatabaseDisplayModel *databaseDisplayModel;
-    DeckListModel *deckModel;
     QTreeView *databaseView;
-
-    QTreeView *deckView;
-    KeySignals deckViewKeySignals;
-    CardInfoFrameWidget *cardInfo;
-    PrintingSelector *printingSelector;
-    SearchLineEdit *searchEdit;
     KeySignals searchKeySignals;
-
-    QLabel *nameLabel;
-    LineEditUnfocusable *nameEdit;
-    QLabel *commentsLabel;
-    QTextEdit *commentsEdit;
-    QLabel *bannerCardLabel;
-    QComboBox *bannerCardComboBox;
-    DeckPreviewDeckTagsDisplayWidget *deckTagsDisplayWidget;
-    QLabel *hashLabel1;
-    LineEditUnfocusable *hashLabel;
-    FilterTreeModel *filterModel;
-    QTreeView *filterView;
-    KeySignals filterViewKeySignals;
-    QWidget *filterBox;
 
     QMenu *deckMenu, *viewMenu, *cardInfoDockMenu, *deckDockMenu, *filterDockMenu, *printingSelectorDockMenu,
         *analyzeDeckMenu, *saveDeckToClipboardMenu, *loadRecentDeckMenu;
@@ -151,8 +115,7 @@ protected:
         *aSaveDeckToClipboard, *aSaveDeckToClipboardNoSetNameAndNumber, *aSaveDeckToClipboardRaw,
         *aSaveDeckToClipboardRawNoSetNameAndNumber, *aPrintDeck, *aExportDeckDecklist, *aAnalyzeDeckDeckstats,
         *aAnalyzeDeckTappedout, *aClose;
-    QAction *aClearFilterAll, *aClearFilterOne;
-    QAction *aAddCard, *aAddCardToSideboard, *aRemoveCard, *aIncrement, *aDecrement, *aSwapCard;
+    QAction *aAddCard, *aAddCardToSideboard;
     QAction *aResetLayout;
     QAction *aCardInfoDockVisible, *aCardInfoDockFloating, *aDeckDockVisible, *aDeckDockFloating, *aFilterDockVisible,
         *aFilterDockFloating, *aPrintingSelectorDockVisible, *aPrintingSelectorDockFloating;
@@ -160,10 +123,6 @@ protected:
     bool modified;
     QVBoxLayout *centralFrame;
     QHBoxLayout *searchLayout;
-    QDockWidget *cardInfoDock;
-    QDockWidget *deckDock;
-    QDockWidget *filterDock;
-    QDockWidget *printingSelectorDock;
     QWidget *centralWidget;
 
 public:
@@ -172,15 +131,18 @@ public:
     [[nodiscard]] virtual QString getTabText() const override = 0; // Subclasses MUST implement this
     void setDeck(DeckLoader *_deckLoader);                         // generic enough
     void setModified(bool _windowModified);                        // generic
+    void setSaveStatus(bool newStatus);                            // generic
     bool confirmClose();                                           // generic
-    virtual void createDeckDock() = 0;
-    virtual void createCardInfoDock() = 0;
-    virtual void createFiltersDock() = 0; // generic
-    virtual void createPrintingSelectorDock() = 0;
     virtual void createMenus() = 0;
     virtual void createCentralFrame() = 0;
     void updateCardInfo(CardInfoPtr _card);
     void addCardHelper(CardInfoPtr info, QString zoneName); // reasonably generic
+    DeckEditorCardInfoDockWidget *cardInfoDockWidget;
+    DeckEditorDeckDockWidget *deckDockWidget;
+    DeckEditorFilterDockWidget *filterDockWidget;
+    DeckEditorPrintingSelectorDockWidget *printingSelectorDockWidget;
+    CardDatabaseDisplayModel *databaseDisplayModel;
+    SearchLineEdit *searchEdit;
 
 public slots:
     void closeRequest(bool forced = false) override; // generic
