+++ cockatrice/src/client/ui/widgets/deck_editor/deck_editor_printing_selector_dock_widget.cpp
@@ -1,0 +1,37 @@
+#include "deck_editor_printing_selector_dock_widget.h"
+
+#include "../../../tabs/tab_generic_deck_editor.h"
+
+#include <QVBoxLayout>
+
+DeckEditorPrintingSelectorDockWidget::DeckEditorPrintingSelectorDockWidget(QWidget *parent,
+                                                                           TabGenericDeckEditor *_deckEditor)
+    : QDockWidget(parent), deckEditor(_deckEditor)
+{
+    setObjectName("printingSelectorDock");
+
+    setAllowedAreas(Qt::LeftDockWidgetArea | Qt::RightDockWidgetArea);
+    setFeatures(QDockWidget::DockWidgetClosable | QDockWidget::DockWidgetFloatable | QDockWidget::DockWidgetMovable);
+    setFloating(false);
+
+    createPrintingSelectorDock();
+}
+
+void DeckEditorPrintingSelectorDockWidget::createPrintingSelectorDock()
+{
+    printingSelector = new PrintingSelector(this, deckEditor);
+    printingSelector->setObjectName("printingSelector");
+    auto *printingSelectorFrame = new QVBoxLayout;
+    printingSelectorFrame->setObjectName("printingSelectorFrame");
+    printingSelectorFrame->addWidget(printingSelector);
+
+    auto *printingSelectorDockContents = new QWidget();
+    printingSelectorDockContents->setObjectName("printingSelectorDockContents");
+    printingSelectorDockContents->setLayout(printingSelectorFrame);
+    setWidget(printingSelectorDockContents);
+
+    installEventFilter(this);
+    // connect(printingSelectorDock, SIGNAL(topLevelChanged(bool)), this, SLOT(dockTopLevelChanged(bool)));
+
+    // addDockWidget(Qt::RightDockWidgetArea, printingSelectorDock);
+}
