+++ cockatrice/src/client/ui/widgets/printing_selector/card_amount_widget.cpp
@@ -16,7 +16,7 @@
  * @param zoneName The zone name (e.g., DECK_ZONE_MAIN or DECK_ZONE_SIDE).
  */
 CardAmountWidget::CardAmountWidget(QWidget *parent,
-                                   TabDeckEditor *deckEditor,
+                                   TabGenericDeckEditor *deckEditor,
                                    DeckListModel *deckModel,
                                    QTreeView *deckView,
                                    QSlider *cardSizeSlider,
