+++ cockatrice/src/client/ui/widgets/printing_selector/printing_selector_card_display_widget.cpp
@@ -28,7 +28,7 @@
  * @param _currentZone The current zone in which the card is located.
  */
 PrintingSelectorCardDisplayWidget::PrintingSelectorCardDisplayWidget(QWidget *parent,
-                                                                     TabDeckEditor *_deckEditor,
+                                                                     TabGenericDeckEditor *_deckEditor,
                                                                      DeckListModel *_deckModel,
                                                                      QTreeView *_deckView,
                                                                      QSlider *_cardSizeSlider,
