+++ cockatrice/src/client/ui/widgets/printing_selector/printing_selector_card_display_widget.h
@@ -1,16 +1,13 @@
 #ifndef PRINTING_SELECTOR_CARD_DISPLAY_WIDGET_H
 #define PRINTING_SELECTOR_CARD_DISPLAY_WIDGET_H
 
-#include "../../../../client/ui/widgets/cards/card_info_picture_widget.h"
 #include "../../../../deck/deck_list_model.h"
 #include "../../../../game/cards/card_database.h"
-#include "../../../tabs/tab_deck_editor.h"
+#include "../../../tabs/tab_generic_deck_editor.h"
 #include "printing_selector_card_overlay_widget.h"
 #include "set_name_and_collectors_number_display_widget.h"
 
 #include <QPainter>
-#include <QTreeView>
-#include <QVBoxLayout>
 #include <QWidget>
 
 class PrintingSelectorCardDisplayWidget : public QWidget
@@ -19,7 +16,7 @@ class PrintingSelectorCardDisplayWidget : public QWidget
 
 public:
     PrintingSelectorCardDisplayWidget(QWidget *parent,
-                                      TabDeckEditor *_deckEditor,
+                                      TabGenericDeckEditor *_deckEditor,
                                       DeckListModel *_deckModel,
                                       QTreeView *_deckView,
                                       QSlider *_cardSizeSlider,
@@ -36,7 +33,7 @@ signals:
 private:
     QVBoxLayout *layout;
     SetNameAndCollectorsNumberDisplayWidget *setNameAndCollectorsNumberDisplayWidget;
-    TabDeckEditor *deckEditor;
+    TabGenericDeckEditor *deckEditor;
     DeckListModel *deckModel;
     QTreeView *deckView;
     QSlider *cardSizeSlider;
